digital-twin/requirements.txt
tree-sitter-language-pack
python-dotenv
GitPython
networkx
sentence-transformers
torch
openai
fastapi[standard]
supabase
uvicorn